+++ app/Controllers/NewsController.php
@@ -1,0 +1,80 @@
+<?php
+namespace App\Controllers;
+use Classes\News;
+use App\Controllers\CommentController;
+use App\Controllers\DB;
+
+class NewsController
+{
+	private static $instance = null;
+
+	private function __construct()
+	{
+
+
+
+	}
+
+	public static function getInstance()
+	{
+		if (null === self::$instance) {
+			$c = __CLASS__;
+			self::$instance = new $c;
+		}
+		return self::$instance;
+	}
+
+	/**
+	 * list all news
+	 */
+	public function listNews()
+	{
+		$db = DB::getInstance();
+		$rows = $db->select('SELECT * FROM `news`');
+
+		$news = [];
+		foreach ($rows as $row) {
+			$n = new News();
+			$news[] = $n->setId($row['id'])
+				->setTitle($row['title'])
+				->setBody($row['body'])
+				->setCreatedAt($row['created_at']);
+		}
+
+		return $news;
+	}
+
+	/**
+	 * add a record in news table
+	 */
+	public function addNews($title, $body)
+	{
+		$db = DB::getInstance();
+		$sql = "INSERT INTO `news` (`title`, `body`, `created_at`) VALUES('" . $title . "','" . $body . "','" . date('Y-m-d') . "')";
+		$db->exec($sql);
+		return $db->lastInsertId($sql);
+	}
+
+	/**
+	 * deletes a news, and also linked comments
+	 */
+	public function deleteNews($id)
+	{
+		$comments = CommentController::getInstance()->listComments();
+		$idsToDelete = [];
+
+		foreach ($comments as $comment) {
+			if ($comment->getNewsId() == $id) {
+				$idsToDelete[] = $comment->getId();
+			}
+		}
+
+		foreach ($idsToDelete as $id) {
+			CommentController::getInstance()->deleteComment($id);
+		}
+
+		$db = DB::getInstance();
+		$sql = "DELETE FROM `news` WHERE `id`=" . $id;
+		return $db->exec($sql);
+	}
+}