+++ index.php
@@ -2,18 +2,20 @@
 require __DIR__ . '/vendor/autoload.php';
 
 define('ROOT', __DIR__);
-require_once(ROOT . '/utils/NewsManager.php');
-require_once(ROOT . '/utils/CommentManager.php');
+use App\Controllers\NewsController;
+use App\Controllers\CommentController;
 
-foreach (NewsManager::getInstance()->listNews() as $news) {
+
+foreach (NewsController::getInstance()->listNews() as $news) {
 	echo ("############ NEWS " . $news->getTitle() . " ############\n");
 	echo ($news->getBody() . "\n");
-	foreach (CommentManager::getInstance()->listComments() as $comment) {
+	foreach (CommentController::getInstance()->listComments() as $comment) {
 		if ($comment->getNewsId() == $news->getId()) {
 			echo ("Comment " . $comment->getId() . " : " . $comment->getBody() . "\n");
 		}
 	}
 }
 
-$commentManager = CommentManager::getInstance();
+
+$commentManager = CommentController::getInstance();
 $c = $commentManager->listComments();